goods/serializers.py
from rest_framework import serializers

from goods.models import Product, Category


class CategorySerializer(serializers.HyperlinkedModelSerializer):

    class Meta:
        model = Category
        fields = ['id', 'name']


class ProductSerializer(serializers.HyperlinkedModelSerializer):
    category = CategorySerializer()

    class Meta:
        model = Product
        fields = ['id', 'title', 'article', 'category', 'price', 'quantity']
        depth = 1
