requirements.txt
Django==4.0.6
django-debug-toolbar==3.5.0
django-otp==1.1.3
django-otp-twilio==1.0.1
djangorestframework==3.13.1
requests==2.28.1
django-environ==0.8.1
